Add files via upload
# INSIGHTS

Standard paper type contributed the highest revenue at 42% followed by gloss at 33% and poster at 25%. Northeast contributed the highest revenue at 33.47% followed by Southest at 27.91%. 

Sales happens to be best for Parch and Posey in 2016 where all the paper types experienced the highest sales. 

EOG Resources was the customer that contributed the highest Revenue with 382k purchase while Nike contributed the less revenue with 390 purchase. 

Earlie Scheusner is the salesperson that contributed the highest revenue at 4.75% while Nakesha Renn contributed 0.21% which is the lowest. 

Sunday happens to be the day with the highest orders with 560k orders. 

There are four advertising channels. However, 59.77% of revenue came in from the customers that used direct channel to purchase. This means marketing the products directly to the customers is paying off.